+++ src/interface.zig
@@ -3,6 +3,7 @@ const position = @import("position.zig");
 const search = @import("search.zig");
 const std = @import("std");
 const types = @import("types.zig");
+const variable = @import("variable.zig");
 
 pub var g_stop = false;
 pub var search_thread: ?std.Thread = null;
@@ -32,29 +33,48 @@ pub const Option = struct {
     default_value: []const u8 = "",
     current_value: []const u8 = "",
     type: []const u8 = "",
-    min: u16 = 0,
-    max: u16 = 0,
+    min: i32 = 0,
+    max: i32 = 0,
     idx: usize = 0, // Order of Option in the OptionsMap
 
-    pub inline fn initCombo(default: []const u8, current: []const u8) Option {
-        return Option{ .type = "combo", .default_value = default, .current_value = current };
+    pub inline fn initCombo(allocator: std.mem.Allocator, default: []const u8, current: []const u8) !Option {
+        return Option{ .type = "combo", .default_value = try allocator.dupe(u8, default), .current_value = try allocator.dupe(u8, current) };
     }
 
-    pub inline fn initSpin(v: []const u8, min: u16, max: u16) Option {
-        return Option{ .type = "spin", .default_value = v, .current_value = v, .max = max, .min = min };
+    pub inline fn initSpin(allocator: std.mem.Allocator, v: []const u8, min: i32, max: i32) !Option {
+        return Option{ .type = "spin", .default_value = try allocator.dupe(u8, v), .current_value = try allocator.dupe(u8, v), .max = max, .min = min };
     }
 
-    pub inline fn initCheck(default: []const u8, current: []const u8) Option {
-        return Option{ .type = "check", .default_value = default, .current_value = current };
+    pub inline fn initCheck(allocator: std.mem.Allocator, default: []const u8, current: []const u8) !Option {
+        return Option{ .type = "check", .default_value = try allocator.dupe(u8, default), .current_value = try allocator.dupe(u8, current) };
     }
 };
 
 pub fn initOptions(allocator: std.mem.Allocator, options: *std.StringArrayHashMapUnmanaged(Option)) !void {
-    try options.put(allocator, "Hash", Option.initSpin("256", 0, 65535));
-    try options.put(allocator, "Threads", Option.initSpin("1", 1, 1));
-    try options.put(allocator, "Evaluation", Option.initCombo("PSQ var PSQ var Shannon", "PSQ"));
-    try options.put(allocator, "Search", Option.initCombo("NegamaxAlphaBeta var NegamaxAlphaBeta var Random", "NegamaxAlphaBeta"));
-    try options.put(allocator, "UCI_Chess960", Option.initCheck("false", "false"));
+    try options.put(allocator, "Hash", try Option.initSpin(allocator, "256", 0, 65535));
+    try options.put(allocator, "Threads", try Option.initSpin(allocator, "1", 1, 1));
+    try options.put(allocator, "Evaluation", try Option.initCombo(allocator, "PSQ var PSQ var Shannon", "PSQ"));
+    try options.put(allocator, "Search", try Option.initCombo(allocator, "NegamaxAlphaBeta var NegamaxAlphaBeta var Random", "NegamaxAlphaBeta"));
+    try options.put(allocator, "UCI_Chess960", try Option.initCheck(allocator, "false", "false"));
+    for (variable.tunables) |tunable| {
+        const min: i32 = @intCast(tunable.min orelse std.math.minInt(types.Value));
+        const max: i32 = @intCast(tunable.max orelse std.math.maxInt(types.Value));
+
+        var buffer: [32]u8 = undefined;
+        const slice = try std.fmt.bufPrint(&buffer, "{d}", .{tunable.default});
+
+        try options.put(allocator, tunable.name, try Option.initSpin(allocator, slice, min, max));
+    }
+}
+
+pub fn deinitOptions(allocator: std.mem.Allocator, options: *std.StringArrayHashMapUnmanaged(Option)) void {
+    const keys = options.keys();
+    for (keys) |key| {
+        const option: Option = options.get(key).?;
+        allocator.free(option.current_value);
+        allocator.free(option.default_value);
+    }
+    options.deinit(allocator);
 }
 
 pub fn printOptions(writer: *std.Io.Writer, options: std.StringArrayHashMapUnmanaged(Option)) void {
@@ -71,7 +91,7 @@ pub fn printOptions(writer: *std.Io.Writer, options: std.StringArrayHashMapUnman
 
 pub fn loop(allocator: std.mem.Allocator, stdin: *std.Io.Reader, stdout: *std.Io.Writer) !void {
     var options: std.StringArrayHashMapUnmanaged(Option) = .empty;
-    defer options.deinit(allocator);
+    defer deinitOptions(allocator, &options);
     try initOptions(allocator, &options);
 
     var states: StateList = .empty;
@@ -277,7 +297,16 @@ fn cmd_setoption(allocator: std.mem.Allocator, tokens: anytype, options: *std.St
             } else if (value_parsed < option.min) {
                 return error.LowerBoundBreached;
             }
+
+            // If option name is tunable edit variable.tunables
+            // tunables are only spin
+            inline for (&variable.tunables) |*tunable| {
+                if (std.ascii.eqlIgnoreCase(tunable.name, name)) {
+                    tunable.default = @intCast(value_parsed);
+                }
+            }
         }
+
         try options.put(allocator, name, option);
     } else {
         return error.UnknownOption;
@@ -336,7 +365,6 @@ fn cmd_go(allocator: std.mem.Allocator, stdout: *std.Io.Writer, noalias pos: *po
     limits.start = types.now();
 
     while (tokens.next()) |token_name| {
-        std.debug.print("token_name {s}\n", .{token_name});
         // Needs to be the last command on the line
         if (std.ascii.eqlIgnoreCase("searchmoves", token_name)) {
             // TODO
@@ -503,7 +531,7 @@ pub fn cmd_bench(allocator: std.mem.Allocator, stdout: *std.Io.Writer) anyerror!
 
     for (list.items) |fen| {
         var options: std.StringArrayHashMapUnmanaged(Option) = .empty;
-        defer options.deinit(allocator);
+        defer deinitOptions(allocator, &options);
         try initOptions(allocator, &options);
 
         var states: StateList = .empty;